Updated navigation in header:
- Updated the links within the dropdown navigation menus
- Updated the icons in those menus to use icon fonts instead of PNG images
- Added a CSS file and icon font files to facilitate new icon setup
- Made some slight changes to CSS in default.css to better match design on developers.mendix.com
- Removed some old CSS in default.css that was previously used to create navigation icons

On possible outstanding item with this.  The links to the Confluence site (world.mendix.com) are using static URLs in header.html.  I know the preference is to use variables in place of the domains, but I don't think there is an existing variable for the Confluence  domain and I couldn't remember how to create a new one.